requirements.txt
# Core packages
streamlit>=1.31.1
flask>=2.3.3
pandas>=2.2.0
numpy>=1.26.2
openai>=1.12.0
python-dotenv>=1.0.0
plotly>=5.18.0
scikit-learn>=1.4.0

# Data processing and analysis
matplotlib>=3.8.2
requests>=2.31.0
scipy>=1.12.0

# NLP packages
transformers>=4.37.0
torch>=2.1.0
nltk>=3.8.1

# For development
pytest>=7.4.3
black>=23.12.0
isort>=5.12.0
flake8>=6.1.0

# huggingface-hub>=0.20.1  # For downloading models from HuggingFace
# sentence-transformers>=2.2.2  # For text embeddings